+++ Epika/Views/SubScreens/RecentExplorationLogsView.swift
@@ -528,7 +528,7 @@ private struct EncounterDetailView: View {
     @State private var battleLogEntries: [BattleLogEntry] = []
     @State private var isLoadingBattleLog = false
     @State private var battleLogError: String?
-    @State private var actorIdentifierToMemberId: [String: Int32] = [:]
+    @State private var actorIdentifierToMemberId: [String: UInt8] = [:]
     @State private var actorIcons: [String: CharacterIconInfo] = [:]
 
     var body: some View {
@@ -675,7 +675,7 @@ private struct EncounterDetailView: View {
         let name = entry.metadata["name"].flatMap { $0.isEmpty ? nil : $0 } ?? "-"
         let level = entry.metadata["level"].flatMap(Int.init)
         let job = entry.metadata["job"].flatMap { $0.isEmpty ? nil : $0 }
-        let memberId = entry.metadata["partyMemberId"].flatMap(Int32.init)
+        let memberId = entry.metadata["partyMemberId"].flatMap(UInt8.init)
 
         return ParticipantState(id: rawId,
                                 name: name,
@@ -721,7 +721,7 @@ private struct EncounterDetailView: View {
         let maxHP: Int
         let level: Int?
         let jobName: String?
-        let partyMemberId: Int32?
+        let partyMemberId: UInt8?
         let role: Role
         let order: Int
     }
@@ -784,7 +784,7 @@ private struct EncounterDetailView: View {
             let archive = try fetchBattleLogArchive()
             battleLogEntries = archive.entries
 
-            var memberMap: [String: Int32] = [:]
+            var memberMap: [String: UInt8] = [:]
             var iconMap: [String: CharacterIconInfo] = [:]
 
             for participant in archive.playerSnapshots {
@@ -835,7 +835,7 @@ private struct EncounterDetailView: View {
         return actorIcons[identifier]
     }
 
-    private func iconInfo(forMember memberId: Int32?) -> CharacterIconInfo? {
+    private func iconInfo(forMember memberId: UInt8?) -> CharacterIconInfo? {
         guard let memberId else { return nil }
         if let identifier = actorIdentifierToMemberId.first(where: { $0.value == memberId })?.key {
             return actorIcons[identifier]
